subgraph service, mocks and tests adjusted. Files moved to proper directories to follow convention
